Mongsil-1 [Refactoring] Modify resource name
- 리소스 네이밍 변경
**layout**
what_where.xml
**strings**
where_description
**drawables**
where_description_size
**ids**
what_where_description
