+++ app/src/main/res/layout/activity_main.xml
@@ -14,7 +14,7 @@ fileSystemWindows:
         tools:context=".ui.MainActivity">
 
         <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/main_fragment"
+            android:id="@+id/fcv_main_fragment_container"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
