+++ app/src/main/res/layout/dialog_emoticon.xml
@@ -12,7 +12,7 @@
         android:background="@color/white"
         android:orientation="vertical">
         <ImageView
-            android:id="@+id/emoticon_btn_close"
+            android:id="@+id/btn_emoticon_close_button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="end"
@@ -25,14 +25,14 @@
             android:textSize="26sp"
             android:textColor="#333333"
             android:layout_marginBottom="@dimen/margin30"
-            android:text="오늘 기분은 어때요?" />
+            android:text="@string/emoticon_title" />
         <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/emoticon_recycler_view"
+            android:id="@+id/rv_emoticon_list"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             android:layout_weight="1"/>
         <TextView
-            android:id="@+id/emoticon_btn_select"
+            android:id="@+id/btn_emoticon_select_button"
             android:layout_width="180dp"
             android:layout_height="56dp"
             android:layout_gravity="center"
@@ -40,7 +40,7 @@
             android:gravity="center"
             android:text="감정 선택"
             android:backgroundTint="@color/colorYellow"
-            android:background="@drawable/datepicker_button_shape"
+            android:background="@drawable/round_button_shape"
             android:textStyle="bold"
             android:textColor="#6d5107"/>
     </LinearLayout>
