+++ app/src/main/res/layout/fragment_calendar.xml
@@ -12,13 +12,13 @@
         android:layout_height="match_parent">
 
         <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/calendar_rv_day_list"
+            android:id="@+id/rv_calendar_day_list"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:visibility="gone"
             android:orientation="vertical"/>
         <com.applandeo.materialcalendarview.CalendarView
-            android:id="@+id/calendar_view"
+            android:id="@+id/cv_calendar_view"
             android:layout_width="match_parent"
             android:layout_height="match_parent"/>
 <!--        <com.prolificinteractive.materialcalendarview.MaterialCalendarView-->
@@ -36,7 +36,7 @@
 <!--            />-->
 
         <com.google.android.material.floatingactionbutton.FloatingActionButton
-            android:id="@+id/floatingActionButton"
+            android:id="@+id/fab_calendar_floatingActionButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             app:layout_constraintBottom_toBottomOf="parent"
