+++ app/src/main/res/layout/fragment_locker.xml
@@ -11,21 +11,21 @@
         android:layout_height="match_parent"
         android:background="@color/purple_200">
         <androidx.appcompat.widget.Toolbar
-            android:id="@+id/locker_toolbar"
+            android:id="@+id/tb_locker"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/locker_recycler_view"
+            app:layout_constraintTop_toBottomOf="@+id/rv_locker_list"
             android:background="@color/purple_500"/>
 
         <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/locker_recycler_view"
+            android:id="@+id/rv_locker_list"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             app:layout_constraintVertical_weight="1"
-            app:layout_constraintTop_toBottomOf="@+id/locker_toolbar"
+            app:layout_constraintTop_toBottomOf="@+id/tb_locker"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"/>
