+++ app/src/main/res/layout/item_emoticon.xml
@@ -19,13 +19,13 @@
         android:elevation="@dimen/margin8"
         android:background="@color/white">
         <com.cashproject.mongsil.util.SquareLayout
-            android:id="@+id/item_emoticon_background"
+            android:id="@+id/iv_emoticon_background"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:orientation="vertical"
             android:gravity="center">
             <ImageView
-                android:id="@+id/item_emoticon_icon"
+                android:id="@+id/iv_emoticon_icon"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="3dp"
@@ -33,16 +33,16 @@
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toTopOf="parent"
                 app:layout_constraintVertical_chainStyle="packed"
-                app:layout_constraintBottom_toTopOf="@id/item_emoticon_emotion"
+                app:layout_constraintBottom_toTopOf="@id/tv_emoticon_emotion"
                 tools:src="@drawable/emoticon_01_happy"/>
             <TextView
-                android:id="@+id/item_emoticon_emotion"
+                android:id="@+id/tv_emoticon_emotion"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/item_emoticon_icon"
+                app:layout_constraintTop_toBottomOf="@+id/iv_emoticon_icon"
                 tools:text = "행복"
                 android:text="@{emoticon.emotion}"/>
         </com.cashproject.mongsil.util.SquareLayout>
