+++ app/src/main/res/layout/item_locker.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
         <ImageView
-            android:id="@+id/item_saying_image"
+            android:id="@+id/iv_locker_image"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="#CDDC39"
