+++ app/src/main/res/layout/item_schedule.xml
@@ -10,7 +10,7 @@
         android:layout_height="50dp">
 
         <TextView
-            android:id="@+id/tv_date"
+            android:id="@+id/tv_schedule_date"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="2dp"
